tasks/errorcollection/requirements.txt
pillow
qrcode
zbarlight
numpy